Instr: Add CachedRunner
The cached runner precomputes sequence of instructions by executing the
Simplicity on the Bit Machine during construction. The final error or
success is saved for later.
The next method pushes cached instructions (from the instruction cache)
into the buffer of past instructions. Both buffers are accessible via
getters.
Tests assert that the cached execution is equal to the uncached one.