Exec: Add Cell
When we display frames, we display the cursor like any other bit.
The Cell struct represents the possible values of a bit / cell.
In the future, we can think about replacing Vec<bool> in Frame with
Vec<Cell>, if it is fast to implement. We can also think about
storing uninitialized bits as Cell::Uninitialized.