Function: Add module
Implement Simplicity in terms of functions and values.
Every step of the evaluation has an expression and an input value.
This can be used to return helpful error messages to the user.
In the future, we could implement a debugger that iterates over
the individual functional steps, which would show users how values
travel through the program.