Util: Rename Expression -> DisplayExpression
This type is used only to print expressions in a flat unshared manner.
We will add this to Node in rust-simplicity as a special way to display
values, similar to SecretKey::display_secret. The lack of sharing means
that the resulting string may be exponentially larger than the original
expression.