Enable run button after animation played
I _think_ the async JS function button_success_animation returns
after it awaited its timeout. This would mean the button is enabled
after the animation played. There is a lot of async going on and I don't
understand all of it yet, but this commit should reduce the potential
button spam somewhat.