Fixed the canvas mouse event blocking
The canvas was blocking mouse events to components that were behind it.
Simplified the previous solution which only partially passed mouse events to the canvas
The canvas now receives no events from the mouse so all events will act normally with components behind it and all mousemove events from the body will be passed onto it so the eyes move with the cursor.