refactor: Remove stateful errors
It is not helpful to display the Simplicity input value when an
assertion fails. After this commit, we just display that an assertion
failed without any additional info. We don't provide any reason why the
assertion failed or how to fix it. This is horrible UX, but the error
messages were not helpful to begin with. Simfony debug symbols will
provide helpful error messages.