feat: Add TxParams
Add the struct TxParams which holds all parameters that we use to
construct a transaction or a transaction environment. Run the examples
unit tests with this environment.

We have to recompute all example signatures. Use the opportunity to
switch the lock times and sequences to values that fit into i32.