chore: Remove program from url parameters
The program text is by far the largest part of the IDE state. There is a
limited on the maximum size of a URL, so it is useful to move the
program out of the URL. Users can copy the program text to the clipboard
using the button that I reintroduce with this commit.