feat: Generate default program with first key
The secret keys were randomized in the previous commit. This commit
adapts the default program to be P2PK for the first of these keys.

The signature data is currently not being generated and it will not
match the program. The program will fail to run, unless the user updates
the signatures.

To ensure that there is a first key, this commit makes the key count
NonZeroU32.