feat: Use local storage
URL parameters were limited in size, which put a limit on the maximum
program size that could be encoded. This commit moves away from stateful
URLs and uses the browser's local storage instead. Every time the Run
button is clicked, the storage is updated. Every time the app is
restarted, it reads from the storage. If the storage is empty, then
there are sane defaults, as before.

As a side effect, this commit disables the Share button. I plan to let
users copy and paste a large blob of characters which will encode the
app's state. This will have to wait until a later commit.